ai-service/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
openai==1.23.2
langchain==0.1.16
langchain-openai==0.1.3
httpx==0.27.0
